kolesa/kolesa_back/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.2
Django==3.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-jwt==1.19.2
pycparser==2.21
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
